+++ openglProject/GameEngine.h
@@ -8,7 +8,7 @@ public:
 	GameEngine();
 	~GameEngine();
 	virtual void start() = 0;
-	virtual void update(int time) = 0;
+	virtual void update(double time) = 0;
 	virtual void draw(Renderer* renderer) = 0;
 };
